changed package name and errors
