+++ authz/providers/azure/azure_test.go
@@ -28,7 +28,7 @@ import (
 	"go.kubeguard.dev/guard/authz/providers/azure/data"
 	authzOpts "go.kubeguard.dev/guard/authz/providers/azure/options"
 	"go.kubeguard.dev/guard/authz/providers/azure/rbac"
-	oputil "go.kubeguard.dev/guard/util/operations"
+	azureutils "go.kubeguard.dev/guard/util/azure"
 
 	"github.com/appscode/pat"
 	"github.com/stretchr/testify/assert"
@@ -61,7 +61,7 @@ func clientSetup(serverUrl, mode string) (*Authorizer, error) {
 		ARMEndPoint: serverUrl + "/arm/",
 	}
 
-	dataActionsMap := map[string][]map[string]map[string]oputil.DataAction{}
+	dataActionsMap := map[string][]map[string]map[string]azureutils.DataAction{}
 
 	c.rbacClient, err = rbac.New(opts, authOpts, &authzInfo, dataActionsMap)
 	if err != nil {
