+++ server/server.go
@@ -28,7 +28,7 @@ import (
 	"go.kubeguard.dev/guard/auth/providers/token"
 	"go.kubeguard.dev/guard/authz/providers/azure"
 	"go.kubeguard.dev/guard/authz/providers/azure/data"
-	oputil "go.kubeguard.dev/guard/util/operations"
+	azureutils "go.kubeguard.dev/guard/util/azure"
 
 	"github.com/appscode/pat"
 	"github.com/prometheus/client_golang/prometheus"
@@ -195,18 +195,18 @@ func (s Server) ListenAndServe() {
 
 			switch s.AuthzRecommendedOptions.Azure.AuthzMode {
 			case "arc":
-				clusterType = oputil.ConnectedClusters
+				clusterType = azureutils.ConnectedClusters
 			case "aks":
-				clusterType = oputil.ManagedClusters
+				clusterType = azureutils.ManagedClusters
 			case "fleet":
-				clusterType = oputil.Fleets
+				clusterType = azureutils.Fleets
 			default:
-				klog.Fatalf("Error in setting authzmode for fetching list of resources")
+				klog.Fatalf("Authzmode %s is not supported for fetching list of resources", s.AuthzRecommendedOptions.Azure.AuthzMode)
 			}
 
-			dataActionsMap, err := fetchListOfResources(clusterType, s.AuthRecommendedOptions.Azure.Environment, s.AuthzRecommendedOptions.Azure.AKSAuthzTokenURL, s.AuthRecommendedOptions.Azure.TenantID)
+			dataActionsMap, err := azureutils.FetchListOfResources(clusterType, s.AuthRecommendedOptions.Azure.Environment, s.AuthzRecommendedOptions.Azure.AKSAuthzTokenURL, s.AuthRecommendedOptions.Azure.TenantID)
 			if err != nil {
-				klog.Fatalf("Failed to create map of data actions. Error:%s", err.Error())
+				klog.Fatalf("Failed to create map of data actions. Error:%s", err)
 			}
 
 			authzhandler.dataActionsMap = dataActionsMap
